pyproject.toml
[tool.poetry]
name = "simula-ligths-api"
version = "0.1.0"
description = ""
authors = ["Thorvald Ballestad", "Henrik Finsberg", "Min RK"]
readme = "README.md"
packages = [{include = "simula_ligths_api"}]

[tool.poetry.scripts]
lights = 'simula_ligths_api.lights:app'

[tool.poetry.dependencies]
python = "^3.9"
numpy = "^1.26.2"
pillow = "^10.1.0"
fastapi = "^0.104.1"
pydantic = "^2.5.2"
uvicorn = "^0.24.0.post1"
typer = "^0.9.0"
playwright = "^1.50.0"


[tool.poetry.group.dev.dependencies]
ruff = "^0.1.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
